requirements.txt
Flask==3.0.2
Flask_Cors==4.0.0
keras==2.13.1
moviepy==1.0.3
nsfw_detector==1.1.1
numpy==1.24.3
opencv_python==4.9.0.80
opencv_python_headless==4.9.0.80
Pillow==10.2.0
pytube==15.0.0
tensorflow==2.13.1
tensorflow_intel==2.13.1
